Initial graph
Prompt: I want you to make a visualization that will be about the difficulty of Geometry Dash levels. I have a blog post at http://localhost:4321/blog/why-geometry-dash-is-the-worlds-hardest-game (don't hardcode, make this a constant so I can change the URL easily) that has a bunch of text in the format of Difficulty value: <a number>. The level's name and author will be in an h3 before the difficulty value. After the difficulty value may be links to YouTube and GDBrowser. After that will be commentary. Find the commentary by taking everything until the next h3 after the links to YouTube/GDBrowser.

Using these heuristics, parse the response and create a vertical bar graph. The x-axis will be the name of the level (will probably have to be rendered vertically to fit since there will be 20+ levels) and the y-axis will be the difficulty value (don't label the y-axis though). Reserve a bit of space on the right side of the screen to show the level title with author, links to YouTube/GDBrowser, and commentary whenever the user clicks on the bar corresponding to a level.

However, when the user first visits the website, it should not show the complete graph. Make it only show the first level with it's commentary opened. Then show forward/back arrows at the bottom right of the screen. Only when the user clicks the forward button should the graph then zoom out just enough to render the difficulty values of the first two levels. After zooming out, display the commentary for the second level automatically. This should go on until the last level. The hardest level will have a value in the millions so it will be expected that the easiest levels will be invisible at the end. This is to instill the viewer with an appreciation of the scale of GD difficulty.