Allow selecting level name on x-axis
Prompt: Weird bug: when resizing the window vertically, the x-axis doesn't move properly. The rest of the chart resizes perfectly but the x-axis is stuck in the same place when the page was first loaded. Also, the user should be able to select by hovering over a level's name in the x-axis as well as the bar.