Improve mobile layout
Prompt: Desktop version is looking perfect layout wise, let's move on to mobile now. I see you made the commentary go to the top instead of the right on mobile which is good. However, it seems like the graph is still being laid out as if it were desktop, taking up the full height (this being partially behind the commentary) and leaving a space on the right for commentary. Make the mobile graph take up the full width, and height wise, it should take up the remainder of whatever height is left after the commentary.